+++ HISTORY.rst
@@ -45,6 +45,7 @@ Unreleased
 
 * Add new version for Odoo 15.0
 * [15.0] Need docker-ce 20 instead of docker-ce 18 for building Odoo 15 on debian:bullseye
+* Publish images on ghcr.io
 
 **Documentation**